+++ scripts/shared/pathsByLanguageVersion.js
